Complete WorkoutDetailsScreen (#3)
* Add WorkoutScreen

* Add images for workouts

* Complete WorkoutDetailsScreen

I added hard coded wokout data in a WorkoutService to load up some
example data while playing around with the UI. I also added a basic
service locator data structure to allow injecting the WorkoutService.

I added Exercise and Workout models with a Workout having a list of
Exercises.

I updated the WorkoutDetailsScreen to use the exercises as items in
the Stepper. Tapping the stepper item navigates to an
ExerciseDetailsScreen.

I added an ExerciseDetailsScreen with a video and a timer.