Listen to messages, keep fresh, update on scroll (#6)
We keep a list of chat messages in the ChatService along with a
StreamController that exposes a stream that the view listens to
and passes the list to a listview.builder.

In the ChatService there are functions to start and stop listening
for new messages, so that when a new message is added to the
firestore, the local lit of messages will be updated and emitted via
the StreamController.

When the last retrieved message is built we retrieve the next 10,
add them to the list and emit the whole list.